[Casts] Change C-style casts (int) and (float) into static_casts

+++ src/_P058_HT16K33_KeyPad.ino
@@ -123,7 +123,7 @@ boolean Plugin_058(uint8_t function, struct EventStruct *event, String& string)
 
         if (P058_data->readKey(key))
         {
-          UserVar[event->BaseVarIndex] = (float)key;
+          UserVar[event->BaseVarIndex] = key;
           event->sensorType            = Sensor_VType::SENSOR_TYPE_SWITCH;
 
           if (loglevelActiveFor(LOG_LEVEL_INFO)) {
